add initial tests for governance modifiable params
add auth, blob, blobstream, and consensus governance param tests

add governance params tests for distribution

add tests for modifying gov params

add tests for modifying ibc params

add tests for modifying slashing params

add tests for modifying staking params

ci: remove version_bump job (#2962)

Closes https://github.com/celestiaorg/celestia-app/issues/2884

- Remove the version_bump and branch_name jobs
- Extract the goreleaser jobs to a new file

Can still create a release. See
https://github.com/rootulp/celestia-app/actions/runs/7324051789/job/19947265102

docs: add column for governance modifiable params in resource_pricing (#2974)

Addresses #2966 by using the governance modifiable params tests in PR
All four of the params are modifiable by governance, which matches the
specs in
[params.md](https://github.com/celestiaorg/celestia-app/blob/main/specs/src/specs/params.md).

- [x] New and updated code has appropriate documentation
- [x] New and updated code has new and/or updated testing
- [x] Required CI checks are passing
- [x] Visual proof for any user facing features like CLI or
documentation updates
- [ ] Linked issues closed with keywords

chore: remove unnecessary conversions (#2975)

Use linter to detect unnecessary type conversions
https://github.com/mdempsky/unconvert

ci: enable govulncheck (#2969)

Closes https://github.com/celestiaorg/celestia-app/issues/2968
Supersedes https://github.com/celestiaorg/celestia-app/pull/2744

- CI workflow is inspired by
https://github.com/celestiaorg/celestia-core/blob/main/.github/workflows/govulncheck.yml
- Bumped to Go 1.21.5 to resolve some vulnerabilities identified by
govulncheck

improve naming of governance params test suite

fix govHandler and remove test for an unmodifiable param.

remove unneeded minter setup

+++ pkg/shares/info_byte.go
@@ -38,6 +38,6 @@ func (i InfoByte) IsSequenceStart() bool {
 
 func ParseInfoByte(i byte) (InfoByte, error) {
 	isSequenceStart := i%2 == 1
-	version := uint8(i) >> 1
+	version := i >> 1
 	return NewInfoByte(version, isSequenceStart)
 }
