+++ x/mint/test/mint_test.go
@@ -37,7 +37,7 @@ func (s *IntegrationTestSuite) SetupSuite() {
 	//
 	// Note: if TimeIotaMs is removed from CometBFT, this technique will no
 	// longer work.
-	cparams.Block.TimeIotaMs = int64(sixMonths.Milliseconds())
+	cparams.Block.TimeIotaMs = sixMonths.Milliseconds()
 
 	cfg := testnode.DefaultConfig().
 		WithConsensusParams(cparams)
